A bug in sudo may allow an attacker to corrupt the heap by specifying a custom prompt.  errata entry #2

+++ regress/lib/libc/regex/debug.c
@@ -1,3 +1,4 @@
+/*	$OpenBSD: debug.c,v 1.2 1995/04/20 22:39:42 cgd Exp $	*/
 /*	$NetBSD: debug.c,v 1.2 1995/04/20 22:39:42 cgd Exp $	*/
 
 #include <stdio.h>
@@ -218,7 +219,7 @@ FILE *d;
 			fprintf(d, ">");
 			break;
 		default:
-			fprintf(d, "!%d(%d)!", OP(*s), opnd);
+			fprintf(d, "!%ld(%ld)!", (long)OP(*s), (long)opnd);
 			break;
 		}
 		if (!done)
